fix(sortable-table): Remove usage of Array.prototype.findIndex
This required the polyfill and was breaking when it wasn't included
(including on the styleguide React section).

Since the tests did not catch this, we decided to remove the babel
polyfill from our test setup and use babel-runtime instead.

babel-runtime does not provide Array.prototype.findIndex, so we used a
version from lodash instead.

[Finishes #97658788]